refactor: separate files for each class
